server/requirements.txt
adium-theme-ubuntu==0.3.4
ccsm==0.9.12.3
cffi==1.11.5
compizconfig-python==0.9.12.3
cx-Oracle==6.2.1
decorator==4.0.6
Django==1.11.1
django-cors-headers==2.2.0
enum34==1.1.6
gevent==1.3b1
greenlet==0.4.13
idna==2.6
ipaddress==1.0.22
ipython==2.4.1
pexpect==4.0.1
ptyprocess==0.5
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.1
python-decouple==3.1
pytz==2018.3
scour==0.32
simplegeneric==0.8.1
six==1.10.0
unity-lens-photos==1.0
virtualenv==15.2.0
